cleaned up the menu.
