+++ Game-On/app/src/main/AndroidManifest.xml
@@ -74,7 +74,7 @@
         </receiver>
 
         <activity android:name=".activities.UserGameActivity"/>
-        
+
     </application>
 
 </manifest>
