+++ Game-On/app/src/main/java/com/example/ttpm/game_on/activities/HomePagerActivity.java
@@ -9,11 +9,16 @@ import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentStatePagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
+import android.view.Menu;
+import android.view.MenuItem;
+import android.widget.Toast;
 
+import com.example.ttpm.game_on.QueryPreferences;
 import com.example.ttpm.game_on.R;
 import com.example.ttpm.game_on.fragments.HostSearchFragment;
 import com.example.ttpm.game_on.fragments.UserSearchFragment;
 import com.example.ttpm.game_on.fragments.UserProfileFragment;
+import com.parse.ParseUser;
 
 public class HomePagerActivity extends AppCompatActivity {
 
@@ -47,5 +52,43 @@ public class HomePagerActivity extends AppCompatActivity {
                 return 3;
             }
         });
+    }
+
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.menu_main, menu);
+        if (QueryPreferences.getStoredSessionId(this) == null) {
+            MenuItem currentSessionMenuItem = menu.findItem(R.id.menu_action_current_session);
+            currentSessionMenuItem.setVisible(false);
+        }
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case R.id.menu_action_log_out:
+                ParseUser currentUser1 = ParseUser.getCurrentUser();
+                String currentuses = currentUser1.getUsername();
+                Toast.makeText(this, currentuses + " has logged out.", Toast.LENGTH_LONG).show();
+                ParseUser.logOut();
+                ParseUser currentUser = ParseUser.getCurrentUser();// this will now be null
+                if (currentUser != null) {
+                    Toast.makeText(this, "Error logging out!", Toast.LENGTH_LONG).show();
+                } else {
+                    Intent gohome = new Intent(this, LoginActivity.class);
+                    startActivity(gohome);
+                    this.finish();
+                }
+                return true;
+            case R.id.menu_action_current_session:
+                Intent intent = SessionActivity.newIntent(this);
+                startActivity(intent);
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
     }
 }
