+++ Game-On/app/src/main/java/com/example/ttpm/game_on/fragments/HostSearchFragment.java
@@ -100,27 +100,6 @@ public class HostSearchFragment extends android.support.v4.app.Fragment
     }
 
     @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.menu_action_log_out:
-                ParseUser currentUser1 = ParseUser.getCurrentUser();
-                String currentuses = currentUser1.getUsername();
-                Toast.makeText(getActivity(), currentuses + " has logged out.", Toast.LENGTH_LONG).show();
-                ParseUser.logOut();
-                ParseUser currentUser = ParseUser.getCurrentUser();// this will now be null
-                if (currentUser != null) {
-                    Toast.makeText(getActivity(), "Error logging out!", Toast.LENGTH_LONG).show();
-                } else {
-                    Intent intent = new Intent(getActivity(), SplashActivity.class);
-                    startActivity(intent);
-                }
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
-        }
-    }
-
-    @Override
     public boolean onQueryTextSubmit(String query) {
         return false;
     }
