+++ Game-On/app/src/main/res/menu/fragment_host_search.xml
@@ -9,15 +9,15 @@
         app:actionViewClass="android.support.v7.widget.SearchView"
         app:showAsAction="ifRoom" />
 
-    <item
-        android:id="@+id/menu_action_settings"
-        android:title="@string/menu_action_settings"
-        android:orderInCategory="100"
-        app:showAsAction="never" />
-    <item
-        android:id="@+id/menu_action_log_out"
-        android:title="@string/menu_action_logout"
-        android:orderInCategory="100"
-        app:showAsAction="never" />
+    <!--<item-->
+        <!--android:id="@+id/menu_action_settings"-->
+        <!--android:title="@string/menu_action_settings"-->
+        <!--android:orderInCategory="100"-->
+        <!--app:showAsAction="never" />-->
+    <!--<item-->
+        <!--android:id="@+id/menu_action_log_out"-->
+        <!--android:title="@string/menu_action_logout"-->
+        <!--android:orderInCategory="100"-->
+        <!--app:showAsAction="never" />-->
 
 </menu>
